don't have a default model
if using OpenAI the user should consciously choose the model and
not just use it blindly with a default. Else this could cause
monetary unexpected losses. So if someone wants to use OpenAI they
must specify the model, also by default llama output won't show
the wrong llm this way and leave that field empty.