default to gemma for alice
Gemma is working much better than mistral so far. Mistral has some
bugs it seems in Candle possibly running long-term or else I have
some bug within my threading of it. Gemma is a mirror of it almost
so it feels more like a Mistral/Candle specific issue. Gemma 2b-it
works really well so it may be fine for now.

Ideally Mixtral Dolphin when I get that working in Candle (WIP).